Add hooks for querying and mutating submission data
The module `src/queries.ts` defines custom hooks that can be used to query and mutate submission data. Under the hood they are built on hooks provided by TanStack Query. The main <IonApp> component is now also wrapped in a <PersistQueryClientProvider> component. TanStack Query requires a Provider wrapper, and this one also performs persistent (as opposed to in-memory) caching. It is connected to a Storage object provided by the @ionic/storage package. The `msw` package is used to mock networks requests in testing.