Replace standalone api functions with singleton object pattern.
This replaces the standalone API request functions with an NmdcServerClient class with equivalent methods. A singleton instance of this class is constructed and exported. This allows an API token to be set on this instance and all subsequent requests will use that token.

These changes also add a new method for calling the `/api/me` endpoint to retrieve the current user.