Add Login, Logout, and Token pages
The Login page has a button that initiates the nmdc-server login using the @capacitor/browser plugin and a new environment variable for the server login route.

The Token page is where the backend will redirect to on successful login with the API token in a query parameter. This page collects that token, places it in persistent storage, and redirects to the home page.

The Logout page clears the token from persistent storage and redirects to the root page.