Wrap application in new StoreProvider
These changes wrap the main application in the new StoreProvider. The TanStack query setup has also been moved to a dedicated component so that it can use the `useStore` hook to access the store while setting up its persistence layer. Finally, the routing has also moved to a separate component where it also uses the `useStore` hook in order to decide where to redirect the root route.

A new AuthRoute component has been added that can be used when registering routes that a user must be logged in (as determined by having an API token in the store) to view.